+++ CoffeeEngine/src/CoffeeEngine/IO/ResourceLoader.h
@@ -76,17 +76,6 @@ namespace Coffee {
 
         static void SetWorkingDirectory(const std::filesystem::path& path) { s_WorkingDirectory = path; }
     private:
-        struct ImportData
-        {
-            UUID uuid;
-            std::filesystem::path originalPath;
-
-            template<typename Archive>
-            void serialize(Archive& archive)
-            {
-                archive(CEREAL_NVP(uuid), CEREAL_NVP(originalPath));
-            }
-        };
         
         static void GenerateImportFile(const std::filesystem::path& path);
         static ImportData GetImportData(const std::filesystem::path& path);
